Componentise LogModal, fix bug with modal flicker on toast timeout, clear modal formdata on submit/close
